requirements.txt
flask==2.3.3
python-dotenv==1.0.0
llama-index==0.9.40
llama-index-llms-openai
llama-index-embeddings-openai
openai==1.3.0
pydantic
pypdf==3.17.1
boto3==1.34.0 